fix attachment query
