+++ build.yaml
@@ -1,6 +1,9 @@
 targets:
   $default:
     builders:
+      moor_generator:
+        options:
+          generate_connect_constructor: true
       json_serializable:
         options:
           explicit_to_json: true
